requirements.txt
Flask
numpy
scikit-learn
torch
torchvision
Pillow
scipy